% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmc_db.r
\name{epmc_db}
\alias{epmc_db}
\title{Retrieve external database entities referenced in a given publication}
\usage{
epmc_db(ext_id = NULL, data_src = "med", db = NULL, n_pages = 20)
}
\arguments{
\item{ext_id}{publication identifier}

\item{data_src}{data source, by default Pubmed/MedLine index will be searched.
Other sources Europe PubMed Central supports are:
\describe{
  \item{agr}{Agricola is a bibliographic database of citations to the
  agricultural literature created by the US National Agricultural Library
  and its co-operators.}
  \item{cba}{Chinese Biological Abstracts}
  \item{ctx}{CiteXplore}
  \item{eth}{EthOs Theses, i.e. PhD theses (British Library)}
  \item{hir}{NHS Evidence}
  \item{med}{PubMed/Medline NLM}
  \item{nbk}{Europe PMC Book metadata}
  \item{pat}{Biological Patents}
  \item{pmc}{PubMed Central}
  }}

\item{db}{restrict the response to a specific database:
\describe{
\item{'CHEBI'}{a database and ontology of chemical entities of biological
    interest \url{http://www.ebi.ac.uk/chebi/}}
 \item{'CHEMBL'}{a database of bioactive drug-like small molecules
    \url{https://www.ebi.ac.uk/chembldb/}}
 \item{'EMBL'}{provides a comprehensive record of the world's nucleotide
    sequencing information \url{http://www.ebi.ac.uk/ena/}}
 \item{'INTACT'}{provides a freely available, open
    source database system and analysis tools for molecular interaction data
    \url{http://www.ebi.ac.uk/intact/}}
 \item{'INTERPRO'}{provides functional analysis of proteins by classifying
    them into families and predicting domains and important sites
    \url{http://www.ebi.ac.uk/interpro/}}
 \item{'OMIM'}{a comprehensive and authoritative compendium of human genes and
    genetic phenotypes \url{http://www.ncbi.nlm.nih.gov/omim}}
 \item{'PDB'}{European resource for the collection,
    organisation and dissemination of data on biological macromolecular
    structures \url{http://www.ebi.ac.uk/pdbe/}}
 \item{'UNIPROT'}{comprehensive and freely accessible
    resource of protein sequence and functional information
 \url{http://www.uniprot.org/}}
 }}

\item{n_pages}{Number of pages to be returned. By default, this function
returns 25 records for each page.}
}
\value{
list of 3 including link count and metadata of cross-references.
}
\description{
This function returns EBI database entities referenced in a publication from
Europe PMC RESTful Web Service.
}
\examples{
  \dontrun{
  epmc_db_details("12368864", db = "uniprot", n_pages = 2)
  }
}

