% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmc_ftxt_book.r
\name{epmc_ftxt_book}
\alias{epmc_ftxt_book}
\title{Fetch Europe PMC books}
\usage{
epmc_ftxt_book(ext_id = NULL)
}
\arguments{
\item{ext_id}{character, publication identifier. All book full texts are accessible
either by the PMID or the 'NBK' book number.}
}
\value{
xml_document
}
\description{
Use this function to retrieve book XML formatted full text for the Open
Access subset of the Europe PMC bookshelf.
}
\examples{
  \dontrun{
  epmc_ftxt_book("NBK32884")
  }
}
