% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmc_search.r
\name{epmc_search}
\alias{epmc_search}
\title{Search Europe PMC publication database}
\usage{
epmc_search(query = NULL, id_list = FALSE, n_pages = 50)
}
\arguments{
\item{query}{search query (character vector). For more information how to
build a search query, see \url{http://europepmc.org/Help}}

\item{id_list}{Should only IDs (e.g. PMID) and sources be retrieved for the
given search terms?}

\item{n_pages}{Number of pages to be returned. By default, this function
returns 25 records for each page.}
}
\value{
List of two, number of hits and the retrieved metadata as data.frame
}
\description{
This is the main function to search
Europe PMC RESTful Web Service (\url{http://europepmc.org/RestfulWebService})
}
\examples{
\dontrun{
#Search articles for 'Gabi-Kat'
my.data <- epmc_search(query='Gabi-Kat')

#Get article metadata by DOI
my.data <- epmc_search(query = 'DOI:10.1007/bf00197367')

#Get article metadata by PubMed ID (PMID)
my.data <- epmc_search(query = 'EXT_ID:22246381')

#Get only PLOS Genetics article with EMBL database references
my.data <- epmc_search(query = 'ISSN:1553-7404 HAS_EMBL:y')
}
}
\seealso{
\url{http://europepmc.org/Help}
}

