% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmc_search.r
\name{epmc_search_}
\alias{epmc_search_}
\title{Get one page of results when searching Europe PubMed Central}
\usage{
epmc_search_(query = NULL, limit = 100, id_list = FALSE,
  synonym = FALSE, page_token = NULL, sort = NULL, ...)
}
\arguments{
\item{query}{character, search query. For more information on how to
build a search query, see \url{http://europepmc.org/Help}}

\item{limit}{integer, limit the number of records you wish to retrieve.
By default, 25 are returned.}

\item{id_list}{logical, should only IDs (e.g. PMID) and sources be retrieved
for the given search terms?}

\item{synonym}{logical, synonym search. If TRUE, synonym terms from MeSH
terminology and the UniProt synonym list are queried, too. Disabled by
default.}

\item{page_token}{cursor marking the page}

\item{sort}{character, sort results by order (\code{asc}, \code{desc}) and
sort field (e.g. \code{CITED}, \code{P_PDATE}), seperated with a blank.
For example, sort results  by times cited in descending order:
\code{sort = "CITED desc"}.}

\item{...}{further params from \code{\link{epmc_search}}}
}
\description{
In general, use \code{\link{epmc_search}} instead. It calls this function, calling all
pages within the defined limit.
}
\seealso{
\link{epmc_search}
}

