% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmc_search_by_doi.r
\name{epmc_search_by_doi_}
\alias{epmc_search_by_doi_}
\title{Search Europe PMC by a DOI name}
\usage{
epmc_search_by_doi_(doi, .pb = NULL, output = NULL)
}
\arguments{
\item{doi}{character vector containing DOI names.}

\item{.pb}{progress bar object}

\item{output}{character, what kind of output should be returned. One of
'parsed', 'id_list' or 'raw' As default, parsed key metadata will be
returned as data.frame. 'id_list' returns a list of IDs and sources. Use
'raw' to get full metadata as list. Please be aware that these lists can
become very large.}
}
\description{
Please use \code{\link{epmc_search_by_doi}} instead. It calls this
method, returning open access status information from all your requests.
}
\examples{
\dontrun{
  epmc_search_by_doi_("10.1159/000479962")
}
}
