% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmc_hits.r
\name{epmc_hits}
\alias{epmc_hits}
\title{Get search result count}
\usage{
epmc_hits(query = NULL, ...)
}
\arguments{
\item{query}{query in the Europe PMC syntax}

\item{...}{add query parameters from `epmc_search()`, e.g. synonym=true}
}
\description{
Search over Europe PMC and retrieve the number of results found
}
\examples{
 \dontrun{
 epmc_hits('abstract:"burkholderia pseudomallei"')
 epmc_hits('AUTHORID:"0000-0002-7635-3473"')
 }
}
\seealso{
\code{\link{epmc_search}}
}

